refactor: change namespace of Tokenizer, Character, Token and TokenType
